+++ src/handlers/mod.rs
@@ -1,0 +1,7 @@
+mod prelude;
+mod utils;
+
+pub mod pictures_in_area;
+pub mod pictures;
+pub mod users;
+pub mod login;
